# docs/swagger.yaml
definitions:
...
  schedule_participant_dtos.InviteToScheduleRequest:
    properties:
      email:
        type: string
      schedule_id:
        type: integer
    type: object
...
  schedule_participant_dtos.ScheduleParticipantResponse:
    properties:
      assign_at:
        type: string
      assign_by:
        type: integer
      id:
        type: integer
      invitation_sent_at:
        type: string
      invitation_status:
        type: string
      response_time:
        type: string
      schedule_id:
        type: integer
      status:
        type: string
      workspace_user_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: Timewise API
  title: timewise-api
  version: "1.0"
paths:
...
  /api/v1/workspace_user/verify-invitation:
    put:
      consumes:
      - application/json
      description: Verify member's request invitation (X-User-Email required, X-Workspace-Id
        required)
      parameters:
      - description: Request body
        in: body
        name: schedule_participant
        required: true
        schema:
          $ref: '#/definitions/schedule_participant_dtos.InviteToScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule_participant_dtos.ScheduleParticipantResponse'
      summary: Verify member's request invitation (X-User-Email required, X-Workspace-Id
        required)
      tags:
      - WorkspaceUser
...
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"